feat: fixed flaws
changed scripts so that there are no passwords in it

+++ config-example/newcluster/setup.sh
@@ -1,19 +1,18 @@
 #!/bin/sh
 
-# Enable SSH
+if [ -z "$1" ]
+then
+	echo "Enter password!"
 
-sudo systemctl enable ssh
+else
 
-# Database Setup
-
-sudo chmod +x startdb.sh
-
-sudo ./startdb.sh
-
-echo "$(cat autostart.sh)" > /etc/rc.local
-
-sudo chmod +x kioskmode.sh
-
-sudo ./kioskmode.sh
-
-sudo reboot
+	# Enable SSH
+	sudo systemctl enable ssh
+	# Database Setup
+	sudo chmod +x startdb.sh
+	sudo ./startdb.sh $1
+	echo "$(cat autostart.sh)" > /etc/rc.local
+	sudo chmod +x kioskmode.sh
+	sudo ./kioskmode.sh
+	sudo reboot	
+fi
